+++ android/src/main/java/io/openim/flutter_openim_sdk/manager/IMManager.java
@@ -36,4 +36,12 @@ public class IMManager extends BaseManager {
     public void getLoginStatus(MethodCall methodCall, MethodChannel.Result result) {
         CommonUtil.runMainThreadReturn(result, Open_im_sdk.getLoginStatus());
     }
+
+    public void wakeUp(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.wakeUp(
+                new OnBaseListener(result),
+                value(methodCall, "operationID")
+        );
+    }
+
 }