+++ ios/Classes/Module/ConversationManager.swift
@@ -20,6 +20,7 @@ public class ConversationManager: BaseServiceManager {
         self["setConversationRecvMessageOpt"] = setConversationRecvMessageOpt
         self["getConversationRecvMessageOpt"] = getConversationRecvMessageOpt
         self["setOneConversationPrivateChat"] = setOneConversationPrivateChat
+        self["deleteConversationMsgFromLocalAndSvr"] = deleteConversationMsgFromLocalAndSvr
     }
     
     func setConversationListener(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
@@ -82,6 +83,10 @@ public class ConversationManager: BaseServiceManager {
     
     func setOneConversationPrivateChat(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
         Open_im_sdkSetOneConversationPrivateChat(BaseCallback(result: result), methodCall[string: "operationID"], methodCall[string: "conversationID"],methodCall[bool: "isPrivate"])
+    }
+    
+    func deleteConversationMsgFromLocalAndSvr(methodCall: FlutterMethodCall, result: @escaping FlutterResult){
+        Open_im_sdkDeleteConversationMsgFromLocalAndSvr(BaseCallback(result: result), methodCall[string: "operationID"], methodCall[string: "conversationID"])
     }
 }