+++ lib/src/enum/conversation_type.dart
@@ -1,4 +1,5 @@
 class ConversationType {
   static const single = 1;
   static const group = 2;
+  static const notification = 4;
 }